# docs/swagger.yaml
components:
  schemas:
    api.ErrorResponse:
      properties:
        error:
          example: validation failed
          type: string
      type: object
...
    api.createWorkflowRequest:
      properties:
        overwrite:
          type: boolean
        workflow:
          $ref: '#/components/schemas/workflow.Workflow'
      type: object
    api.createWorkflowResponse:
      properties:
        created_at:
          type: string
        id:
          type: string
        name:
          type: string
        version:
          type: string
      type: object
    api.validateRequest:
      properties:
        dataset_id:
          type: string
      type: object
...
    dataset.ValidationResult:
      properties:
        missing:
          items:
            type: string
          type: array
          uniqueItems: false
        ok:
          type: boolean
        warnings:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
    workflow.DataSchema:
      properties:
        pools:
          additionalProperties:
            $ref: '#/components/schemas/workflow.PoolSchema'
          type: object
      type: object
    workflow.DelaySpec:
      properties:
        max_ms:
          type: integer
        min_ms:
          type: integer
      type: object
    workflow.PoolSchema:
      properties:
        fields:
          items:
            type: string
          type: array
          uniqueItems: false
        min_size:
          type: integer
        optional:
          type: boolean
      type: object
    workflow.Workflow:
      properties:
        base_url:
          type: string
        created_at:
          type: string
        data_schema:
          $ref: '#/components/schemas/workflow.DataSchema'
        default_delay:
          $ref: '#/components/schemas/workflow.DelaySpec'
        description:
          type: string
        estimated_rps_per_vu:
          type: integer
        id:
          type: string
        name:
          type: string
        root:
          items:
            type: integer
          type: array
          uniqueItems: false
        targets:
          items:
            type: string
          type: array
          uniqueItems: false
        thresholds:
          additionalProperties:
            items:
              type: string
            type: array
          type: object
        version:
          type: string
      type: object
info:
  description: 'Zeus is the execution plane for faults-lab. Manteion issues

    commands to it. This document is the integration contract between

    the two. Run lifecycle state machine and cross-cutting concerns

    (dataset handshake, error envelope) are documented in

    docs/api-contract.md.'
  title: Zeus Load Generation API
  version: '1.0'
openapi: 3.1.0
paths:
...
  /workflows:
    get:
      description: "Returns an envelope { \"workflows\": [...] } where each entry\
        \ is a compact\nsummary (id, name, version, targets) \u2014 not the full DSL\
        \ document. Use GET\n/workflows/{id} for the full document."
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: envelope with workflows array
      summary: List workflows
      tags:
      - workflows
    post:
      description: 'Persist a Zeus DSL v2 workflow. The body envelope wraps the workflow
        object

        alongside an overwrite flag; when overwrite is true, an existing workflow
        with

        the same name is replaced. Schema validation runs before persistence; failures

        return 400 with the validator error string. Name conflicts without overwrite

        yield 409.'
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.createWorkflowRequest'
                description: workflow envelope
                summary: body
        description: workflow envelope
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.createWorkflowResponse'
          description: Created
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: invalid JSON or workflow validation error
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: workflow name conflict (overwrite != true)
      summary: Register workflow
      tags:
      - workflows
  /workflows/{id}:
    delete:
      description: 'Refuses with 409 when any non-terminal run is still bound to the
        workflow;

        callers should stop those runs first.'
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '204':
          description: workflow deleted
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: workflow not found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: workflow has active runs
      summary: Delete workflow
      tags:
      - workflows
    get:
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workflow.Workflow'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: workflow not found
      summary: Get workflow
      tags:
      - workflows
...
  /workflows/{id}/validate:
    post:
      description: 'Runs ValidateDataset against the workflow''s DataSchema. Workflows
        without a

        DataSchema short-circuit to {ok: true, warnings: []}. The result body is a

        dataset.ValidationResult. Returns 422 (not 400) when the dataset fails schema

        validation, mirroring run-create''s rejected semantics. 400 is reserved for

        malformed request bodies.'
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.validateRequest'
                description: validation request (dataset id)
                summary: body
        description: validation request (dataset id)
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dataset.ValidationResult'
          description: validation succeeded
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: invalid JSON
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: workflow or dataset not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dataset.ValidationResult'
          description: dataset failed schema validation
      summary: Validate workflow against dataset
      tags:
      - workflows
servers:
- description: Local dev (HTTP)
  url: http://localhost:8080/api/v1
- description: Local dev (HTTPS)
  url: https://localhost:8080/api/v1